split codingame game file
