+++ src/main/scala/com/truelaurel/codingame/architecture/GameResult.scala
@@ -1,0 +1,14 @@
+package com.truelaurel.codingame.architecture
+
+sealed trait GameResult
+
+
+case object Draw extends GameResult
+
+case object WinKO extends GameResult
+
+case object LossKO extends GameResult
+
+case object WinTech extends GameResult
+
+case object LossTech extends GameResult
